cucumber-expressions: Retain position of optional groups (#1076)
When using a regular expression with multiple optional capturing groups. Absent
groups (i.e. a null value) were filtered out. This makes optional
capture groups hard to use:

For example given a parameter type with optional capturing groups:

```java
   parameterTypeRegistry.defineParameterType(new ParameterType<>(
                "textAndOrNumber",
                singletonList("([A-Z]+)?(?: )?([0-9]+)?"),
                new TypeReference<List<String>>() {
                }.getType(),
                new CaptureGroupTransformer<List<String>>() {
                    @Override
                    public List<String> transform(String... args) {
                        return ...;
                    }
                },
                false,
                false)
        );
```

Now the intuition would be that `args[0]` always contains letters and `args[1]`
always contains numbers.

So when we match the expression `{textAndOrNumber}` against `TLA 123`. Then
`args = ["TLA", "123"]`. However when matched `TLA` or `123`
then `args = ["TLA"]` or `args = ["123"]`. This makes it hard to
distinguish between the two.

By not removing the null values we resolve this problem. So when matching
`123` we end up with `args = [null, "123"]`.

However. This will break instances where people were depend on the strict
either-or behaviour. As we did, for example our build in `{string}` type.

Fixes: #511
Fixes: #952