Parse negative numbers in Norwegian (and 59 other languages)
The DecimalFormatSymbols for Norwegian and 59 other languages use the
minus-sign (unicode 8722) instead of the hyphen-minus sign (ascii 45).

While technically correct, Gherkin is written on regular keyboards and
there is no practical way to write a minus-sign. By patching the
`DecimalFormatSymbols` with a regular minus sign we solve this problem.

Additionally, for the same reason, the non-breaking space (ascii 160)
and right single quotation mark (unicode 8217) for thousands separators
are also patched with either a period or colon.

Fixes: #287